implemented reqChunks and delChunks
meant to request and delete chunks that should or should not be loaded.

player cannot move and load chunks though

+++ src/client/models/block/Block.cpp
@@ -22,7 +22,7 @@ namespace block {
 
     Block::Block() {}
 
-    void Block::render(sf::RenderWindow *window, int *chunk, int *plrCoords) {
+    void Block::render(sf::RenderWindow *window, int *chunk, float *plrCoords) {
         this->setPosition(
                 (((float)chunk[0] * CHUNK_SIZE) + this->coords[0] + plrCoords[0]) * BLOCK_SIZE,
                 (((float)chunk[1] * CHUNK_SIZE) + this->coords[1] + plrCoords[1]) * BLOCK_SIZE
