+++ src/client/models/block/Block.h
@@ -6,6 +6,7 @@
 
 #include <nlohmann/json.hpp>
 #include <SFML/Graphics.hpp>
+#include "fmt/core.h"
 #include "../../globvars/globvars.h"
 
 using json = nlohmann::json;
@@ -26,7 +27,7 @@ namespace block {
 
         Block(); // blank constructor so we can make block arrays
 
-        void render(sf::RenderWindow *window, int chunk[2], int plrCoords[2]); // renders the block to the screen
+        void render(sf::RenderWindow *window, int chunk[2], float plrCoords[2]); // renders the block to the screen
 
         // getters
         char* getCoords();
